+++ packages/sanity-webhook-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @last-rev/sanity-webhook-parser
 
+## 0.3.0
+
+### Minor Changes
+
+- 4a3d88c: Major update to Sanity path for libraries. Sanity object structure will now be preserved instead of mapping to Contentful objects.
+
 ## 0.2.1
 
 ### Patch Changes